backend/requirements.txt
fastapi>=0.110.0
uvicorn
sqlmodel
psycopg2
redis
openai
tiktoken  
httpx  
python-jose[cryptography]  
passlib[bcrypt]  
pyjwt 
GitPython
